pyproject.toml
[tool.poetry]
name = "aiven-site-monitor"
version = "0.1.0"
description = ""
authors = ["a.letenkov <a.letenkov@corp.vk.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.1"
pytest = "^6.2.4"
kafka-python = "^2.0.2"
pydantic = "^1.8.2"
aiohttp = "^3.7.4"
pyaml = "^21.8.3"
pytz = "^2021.1"
pytest-mock = "^3.6.1"
unittest = "^0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
